safari-534.53-branch/LayoutTests/java/lc3/JavaObject/JavaObjectToByte-006-expected.txt

LiveConnect 3.0 JavaScript to Java Data Type Conversion JavaObject to byte

PASS dt.PUB_DOUBLE_REPRESENTATION = -128.6;dt.setByte( dt ); dt.PUB_BYTE is -128
PASS dt.PUB_DOUBLE_REPRESENTATION = -128.6;dt.setByte( dt ); dt.getByte() is -128
PASS typeof dt.getByte() is 'number'
PASS successfullyParsed is true

TEST COMPLETE
 
